+++ common/src/main/java/com/wynntils/models/containers/containers/ContentBookContainer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © Wynntils 2024.
+ * Copyright © Wynntils 2024-2025.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
 package com.wynntils.models.containers.containers;
@@ -33,12 +33,12 @@ public class ContentBookContainer extends Container
     }
 
     @Override
-    public int getNextItemSlot() {
+    public short getNextItemSlot() {
         return 69;
     }
 
     @Override
-    public int getPreviousItemSlot() {
+    public short getPreviousItemSlot() {
         return 65;
     }
