+++ common/src/main/java/com/wynntils/models/containers/containers/LeaderboardRewardsContainer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © Wynntils 2024.
+ * Copyright © Wynntils 2024-2025.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
 package com.wynntils.models.containers.containers;
@@ -28,12 +28,12 @@ public class LeaderboardRewardsContainer extends Container implements Scrollable
     }
 
     @Override
-    public int getNextItemSlot() {
+    public short getNextItemSlot() {
         return 23;
     }
 
     @Override
-    public int getPreviousItemSlot() {
+    public short getPreviousItemSlot() {
         return 5;
     }
 }
